+++ config/config.exs
@@ -3,8 +3,6 @@ import Config
 
 config :elixir, :time_zone_database, Tz.TimeZoneDatabase
 
-config :tesla, adapter: {Tesla.Adapter.Hackney, [recv_timeout: 10_000]}
-
 config :inngest,
   # app_host: "https://79c9-99-123-3-121.ngrok-free.app",
   # signing_key: "",